refactor(drawPolygon): move functions to own files
